// src/forum/tags/interfaces/tag.response.interface.ts
import { IAuditFields } from "src/common/interfaces/common.interface";

export interface ITagResponse extends IAuditFields {
    id: string;
    title: string;
    description: string;
    slug: string;
    userId: string;
    username: string;
    nickname: string;
    author: string;
    userRole: string;
    email: string;
    status: boolean;
}